requirements.txt
asgiref==3.7.2
Django==4.2.7
django-filter==23.3
djangorestframework==3.14.0
Markdown==3.5.1
mysqlclient==2.2.0
Pillow==10.1.0
python-dateutil==2.8.2
pytils==0.4.1
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
